+++ src/stores/types/stores.ts
@@ -1,0 +1,12 @@
+import { Method } from 'axios';
+
+export type EndpointType = {
+  url: string;
+  method?: Method;
+};
+
+export type EndpointsType = Record<string, EndpointType> & {
+  auth: EndpointType;
+  getUser?: EndpointType;
+  flag?: EndpointType;
+};