+++ src/utils/logError.ts
@@ -1,0 +1,6 @@
+export default function logError(error: string): void {
+  if (window.is_dev) {
+    // eslint-disable-next-line
+    console.error(`[@ktsstudio/mediaproject-stores] ${error}`);
+  }
+}