+++ CHANGELOG.md
@@ -1,1 +1,6 @@
+### v1.1.0
+- [*] refactor according to api schema
+- [*] fix send user flag method payload 
+
 ### v1.0.0
+- [+] BaseRootStore, BaseSubstore, BaseUserStore, types